Refactor: Consolidate KeyPairRepository tests and add PlayIntegrityRepository DI
- Merged `EcKeyPairRepositoryImplTest` and `RsaKeyPairRepositoryImplTest` into a single `KeyPairRepositoryImplTest`.
- Updated DI module to provide `PlayIntegrityRepository`.
- Removed unused `android.util.Log` import from `KeyAttestationViewModel`.
