+++ android/app/src/main/java/dev/keiji/deviceintegrity/di/RepositoryModule.kt
@@ -9,7 +9,9 @@ import dagger.hilt.components.SingletonComponent
 import dev.keiji.deviceintegrity.repository.contract.PreferencesRepository
 import dev.keiji.deviceintegrity.repository.impl.PreferencesRepositoryImpl
 import dev.keiji.deviceintegrity.repository.contract.KeyPairRepository
+import dev.keiji.deviceintegrity.repository.contract.PlayIntegrityRepository
 import dev.keiji.deviceintegrity.repository.impl.KeyPairRepositoryImpl
+import dev.keiji.deviceintegrity.repository.impl.PlayIntegrityRepositoryImpl
 import kotlinx.coroutines.Dispatchers
 import javax.inject.Singleton
