--- android/repository/impl/src/androidTest/java/dev/keiji/deviceintegrity/repository/impl/EcKeyPairRepositoryImplTest.kt
@@ -1,233 +1,0 @@
-package dev.keiji.deviceintegrity.repository.impl
-
-import android.content.Context
-import android.os.Build
-import android.security.keystore.KeyGenParameterSpec
-import android.security.keystore.KeyProperties
-import androidx.test.core.app.ApplicationProvider
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.test.StandardTestDispatcher
-import kotlinx.coroutines.test.TestCoroutineScheduler
-import kotlinx.coroutines.test.runTest
-import org.junit.After
-import org.junit.Assert.*
-import org.junit.Before
-import org.junit.Test
-import org.junit.runner.RunWith
-import java.security.KeyPairGenerator
-import java.security.KeyStore
-
-@ExperimentalCoroutinesApi
-@RunWith(AndroidJUnit4::class)
-class EcKeyPairRepositoryImplTest {
-
-    private lateinit var context: Context
-    private lateinit var keyStore: KeyStore
-    private lateinit var keyPairRepository: EcKeyPairRepositoryImpl
-
-    // Using Dispatchers.IO directly as it's an instrumentation test.
-    // For more control, a TestCoroutineDispatcher setup for instrumentation tests would be needed.
-    // Let's use a StandardTestDispatcher for runTest.
-    private val testScheduler = TestCoroutineScheduler()
-    private val testDispatcher = StandardTestDispatcher(testScheduler)
-
-
-    private val testAliasBase = "instrumentationTestAlias"
-
-    @Before
-    fun setUp() {
-        context = ApplicationProvider.getApplicationContext<Context>()
-        keyStore = KeyStore.getInstance("AndroidKeyStore")
-        keyStore.load(null)
-
-        // It's important that EcKeyPairRepositoryImpl uses Dispatchers.IO or a dispatcher we can control.
-        // The Impl is hardcoded to use a dispatcher passed in constructor.
-        // For this test, we pass our testDispatcher.
-        keyPairRepository = EcKeyPairRepositoryImpl(testDispatcher, context)
-
-        // Clean up any aliases from previous runs to ensure test isolation
-        cleanupTestKeys()
-    }
-
-    @After
-    fun tearDown() {
-        cleanupTestKeys()
-    }
-
-    private fun cleanupTestKeys() {
-        // Clean up a few potential aliases to avoid issues between test runs
-        for (i in 0..5) {
-            val alias = "$testAliasBase$i"
-            if (keyStore.containsAlias(alias)) {
-                keyStore.deleteEntry(alias)
-            }
-        }
-        // Clean up dynamically generated UUID aliases if possible, though harder to track
-        // This cleanup is mostly for fixed test aliases.
-        // Also, try to clean up any aliases that might have been generated by generateKeyPair tests
-        keyStore.aliases().toList().forEach { alias ->
-            if (alias.startsWith(testAliasBase) || alias.length == 36) { // 36 is UUID length
-                try {
-                    keyStore.deleteEntry(alias)
-                } catch (e: Exception) {
-                    // Ignore errors during cleanup
-                }
-            }
-        }
-    }
-
-    private fun generateTestKey(alias: String): java.security.KeyPair {
-        val keyPairGenerator = KeyPairGenerator.getInstance(KeyProperties.KEY_ALGORITHM_EC, "AndroidKeyStore")
-        keyPairGenerator.initialize(
-            KeyGenParameterSpec.Builder(alias, KeyProperties.PURPOSE_SIGN or KeyProperties.PURPOSE_VERIFY)
-                .setDigests(KeyProperties.DIGEST_SHA256)
-                .setKeySize(256) // Using the known key size
-                .build()
-        )
-        return keyPairGenerator.generateKeyPair()
-    }
-
-    @Test
-    fun getKeyPair_returnsKeyPair_whenAliasExists() = runTest(testScheduler) {
-        val alias = "${testAliasBase}GetKey"
-        val expectedKeyPair = generateTestKey(alias)
-
-        val result = keyPairRepository.getKeyPair(alias)
-
-        assertNotNull(result)
-        assertEquals(
-            expectedKeyPair.public.encoded.toHexString(),
-            result?.public?.encoded?.toHexString()
-        )
-    }
-
-    @Test
-    fun getKeyPair_returnsNull_whenAliasDoesNotExist() = runTest(testScheduler) {
-        val result = keyPairRepository.getKeyPair("nonExistentAliasForGetKey")
-        assertNull(result)
-    }
-
-    @Test
-    fun removeKeyPair_deletesExistingKey() = runTest(testScheduler) {
-        val alias = "${testAliasBase}RemoveKey"
-        generateTestKey(alias)
-        assertTrue(keyStore.containsAlias(alias))
-
-        keyPairRepository.removeKeyPair(alias)
-
-        assertFalse(keyStore.containsAlias(alias))
-    }
-
-    @Test
-    fun removeKeyPair_doesNothing_whenAliasDoesNotExist() = runTest(testScheduler) {
-        val alias = "nonExistentAliasForRemove"
-        assertFalse(keyStore.containsAlias(alias))
-
-        keyPairRepository.removeKeyPair(alias) // Should not throw
-
-        assertFalse(keyStore.containsAlias(alias))
-    }
-
-    @Test
-    fun generateKeyPair_successfullyGeneratesKey_andReturnsKeyPairDataWithCertificates() = runTest(testScheduler) {
-        val challenge = "test_challenge_instrumentation".toByteArray()
-
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.N) {
-            // This success path test expects to run on N+ because of the challenge requirement.
-            // The method EcKeyPairRepositoryImpl.generateKeyPair throws if SDK < N.
-            // So, if we are on pre-N, this test should not proceed to call generateKeyPair.
-            // The exception for pre-N is tested in another dedicated test.
-            println("Skipping generateKeyPair_successfullyGeneratesKey test on pre-N SDK (${Build.VERSION.SDK_INT}) as it requires challenge.")
-            return@runTest
-        }
-
-        val result = keyPairRepository.generateKeyPair(challenge)
-
-        assertNotNull(result)
-        assertTrue(result.keyAlias.isNotEmpty())
-        assertNotNull(result.certificates)
-        assertTrue(result.certificates.isNotEmpty())
-        assertTrue(keyStore.containsAlias(result.keyAlias))
-        assertEquals(result.certificates[0].publicKey, keyStore.getCertificate(result.keyAlias).publicKey)
-
-        // Important: Clean up dynamically generated alias
-        if (keyStore.containsAlias(result.keyAlias)) {
-            keyStore.deleteEntry(result.keyAlias)
-        }
-    }
-
-    @Test
-    fun generateKeyPair_handlesAliasCollision_andRetries() = runTest(testScheduler) {
-        val challenge = "test_challenge_collision".toByteArray()
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.N) {
-            println("Skipping generateKeyPair_handlesAliasCollision test on pre-N SDK (${Build.VERSION.SDK_INT}) as it requires challenge.")
-            return@runTest
-        }
-
-        // Generate one key
-        val keyPairData1 = keyPairRepository.generateKeyPair(challenge)
-        assertNotNull(keyPairData1)
-        assertTrue(keyStore.containsAlias(keyPairData1.keyAlias))
-
-        // Generate another key - should have a different alias due to UUID and retry
-        val keyPairData2 = keyPairRepository.generateKeyPair(challenge)
-        assertNotNull(keyPairData2)
-        assertTrue(keyStore.containsAlias(keyPairData2.keyAlias))
-
-        assertNotEquals(keyPairData1.keyAlias, keyPairData2.keyAlias)
-
-        // Cleanup
-        keyStore.deleteEntry(keyPairData1.keyAlias)
-        keyStore.deleteEntry(keyPairData2.keyAlias)
-    }
-
-    @Test
-    fun generateKeyPair_throwsUnsupportedOperationException_onPreNSDK_whenChallengeProvided() = runTest(testScheduler) {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-            println("Skipping generateKeyPair_throwsUnsupportedOperationException_onPreNSDK test on N+ SDK (${Build.VERSION.SDK_INT}).")
-            return@runTest
-        }
-
-        val challenge = "test_challenge_pre_N".toByteArray() // Challenge content doesn't matter for this exception
-        var exceptionThrown = false
-        try {
-            keyPairRepository.generateKeyPair(challenge)
-        } catch (e: UnsupportedOperationException) {
-            exceptionThrown = true
-            assertTrue(e.message?.contains("not supported before Android N") == true)
-        }
-        assertTrue("UnsupportedOperationException was expected but not thrown on SDK ${Build.VERSION.SDK_INT}.",exceptionThrown)
-    }
-
-    @Test
-    fun generateKeyPair_throwsIllegalArgumentException_onNPlusSDK_whenChallengeIsEmpty() = runTest(testScheduler) {
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.N) {
-            println("Skipping generateKeyPair_throwsIllegalArgumentException_onNPlusSDK test on pre-N SDK (${Build.VERSION.SDK_INT}).")
-            return@runTest
-        }
-
-        val emptyChallenge = byteArrayOf()
-        var exceptionThrown = false
-        try {
-            keyPairRepository.generateKeyPair(emptyChallenge)
-        } catch (e: IllegalArgumentException) {
-            exceptionThrown = true
-            assertTrue(e.message?.contains("Challenge cannot be empty") == true)
-        }
-        assertTrue("IllegalArgumentException was expected but not thrown on SDK ${Build.VERSION.SDK_INT}.",exceptionThrown)
-    }
-
-    @Test
-    fun generateKeyPair_throwsIllegalStateException_whenCertificateChainIsNull() = runTest(testScheduler) {
-        // This specific scenario (AndroidKeyStore successfully generates a key but returns a null chain)
-        // is highly unlikely with the real AndroidKeyStore if key generation succeeded.
-        // The AndroidKeyStore either provides a valid chain or key generation itself fails.
-        // Thus, directly testing the explicit "throw IllegalStateException("Failed to retrieve certificate chain...")"
-        // is very difficult in an instrumentation test without manipulating KeyStore internals.
-        // We trust that if such an inconsistent state *could* occur, our code would throw.
-        // For now, we acknowledge this test is hard to implement reliably for this specific throw.
-        // A successful generation (tested elsewhere) implies a non-null chain.
-        assertTrue("Test for null certificate chain is hard to implement in instrumentation test.", true)
-    }
-}
